requirements.txt
requests
python-dotenv
huggingface-hub[inference]
transformers>=4.37.0
torch
sentencepiece
flask>=2.3.0
lean-dojo>=1.8.0